requirements.txt
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.5
langchain-community==0.0.13
faiss-cpu==1.7.4
pypdf==3.17.4
python-dotenv==1.0.0
tiktoken==0.5.2